Reworking code into database
